Extract ConfigValidator class from Config class
Do configuration validation in a new class.
