+++ lib/rubocop.rb
@@ -587,6 +587,7 @@ require_relative 'rubocop/config'
 require_relative 'rubocop/config_loader_resolver'
 require_relative 'rubocop/config_loader'
 require_relative 'rubocop/config_store'
+require_relative 'rubocop/config_validator'
 require_relative 'rubocop/target_finder'
 require_relative 'rubocop/token'
 require_relative 'rubocop/comment_config'
