+++ spec/rubocop/processed_source_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe RuboCop::ProcessedSource do
 
   subject(:processed_source) { described_class.new(source, ruby_version, path) }
 
-  let(:ruby_version) { RuboCop::Config::KNOWN_RUBIES.last }
+  let(:ruby_version) { RuboCop::ConfigValidator::KNOWN_RUBIES.last }
 
   let(:source) { <<~RUBY }
     # an awesome method
